+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - `Query` and `Body` parameters do optionally accept a factory function to create the instance.
+- `registerControllersFromFolder` function that loads and registers controllers from a folder.
 
 ### Changed
 - Updated `typings` to 1.x version: `ambient` dependencies are new `global` dependencies.